WIP: Add `types.ts` for public exports
